BAP-21019: Switch to PSR-12 coding standard except explicit constants visibility requirement (#35870)
- used PSR-12 except required visibility for constants
- updated php-cs-fixer config
- added doc block comments for changed classes
- updated php-semver-checker

+++ src/Oro/Bundle/ChannelBundle/EventListener/AccountLifetimeListener.php
@@ -72,7 +72,7 @@ class AccountLifetimeListener implements ServiceSubscriberInterface
         foreach ($this->accounts as $account) {
             $lifetimeAmountQb->setParameter('account', $account->getId());
 
-            $lifetimeAmount = (double)$lifetimeAmountQb->getQuery()->getSingleScalarResult();
+            $lifetimeAmount = (float)$lifetimeAmountQb->getQuery()->getSingleScalarResult();
 
             $history = new LifetimeValueHistory();
             $history->setAmount($lifetimeAmount);
